NMS-17739: Issues in Alarm and Event DAO (Null reference and BigInteger conversion) (#7661)
* NMS-17739, initial commit

* NMS-17739, added test case in alarm dao

* NMS-17739, used parameterized query, added some more asserts

* NMS-17739, update after testing

* NMS-17739, added test cases for events

* NMS-17739, typo fixes

* NMS-17739, code review updates for Alarm dao and test cases

* NMS-17739, updated event dao and test cases

+++ opennms-dao-mock/src/main/java/org/opennms/netmgt/dao/mock/MockAlarmDao.java
@@ -82,7 +82,7 @@ public class MockAlarmDao extends AbstractMockDao<OnmsAlarm, Integer> implements
     }
 
     @Override
-    public int getNumAlarmsLastHours(int hours) {
+    public long getNumAlarmsLastHours(int hours) {
         return 0;
     }
